# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Airbnb Clone API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/host/listings": {
      "get": {
        "tags": [
          "Host"
        ],
        "summary": "List Host Listings",
        "description": "List all non-deleted listings owned by the current host.",
        "operationId": "list_host_listings_api_host_listings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HostListingResponse"
                  },
                  "type": "array",
                  "title": "Response List Host Listings Api Host Listings Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Host"
        ],
        "summary": "Create Host Listing",
        "description": "Create a new listing.",
        "operationId": "create_host_listing_api_host_listings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/host/listings/{listing_id}": {
      "get": {
        "tags": [
          "Host"
        ],
        "summary": "Get Host Listing Endpoint",
        "description": "Get a single host-owned listing.",
        "operationId": "get_host_listing_endpoint_api_host_listings__listing_id__get",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Listing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Host"
        ],
        "summary": "Update Host Listing",
        "description": "Full or partial update to a host-owned listing.",
        "operationId": "update_host_listing_api_host_listings__listing_id__put",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Listing Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Host"
        ],
        "summary": "Delete Host Listing",
        "description": "Soft-delete a listing (409 if upcoming bookings exist).",
        "operationId": "delete_host_listing_api_host_listings__listing_id__delete",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Listing Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/host/reservations": {
      "get": {
        "tags": [
          "Host"
        ],
        "summary": "Host Reservations",
        "description": "List all reservations across host's listings, optionally filtered.",
        "operationId": "host_reservations_api_host_reservations_get",
        "parameters": [
          {
            "name": "tab",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(all|upcoming|past)$",
              "default": "all",
              "title": "Tab"
            }
          },
          {
            "name": "listing_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Listing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/uploads": {
      "post": {
        "tags": [
          "Host"
        ],
        "summary": "Upload Image",
        "description": "Upload a listing photo. Validates via magic bytes (not just extension/MIME header).\nMax 5MB. Returns the URL path for use in listing creation.",
        "operationId": "upload_image_api_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List Notifications",
        "description": "List user's notifications, newest first.",
        "operationId": "list_notifications_api_notifications_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark Notification Read",
        "description": "Mark a single notification as read.",
        "operationId": "mark_notification_read_api_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_image_api_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_uploads_post"
      },
...
      "CreateListingRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "property_type": {
            "type": "string",
            "title": "Property Type"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "max_guests": {
            "type": "integer",
            "title": "Max Guests"
          },
          "bedrooms": {
            "type": "integer",
            "title": "Bedrooms"
          },
          "beds": {
            "type": "integer",
            "title": "Beds"
          },
          "baths": {
            "type": "number",
            "title": "Baths"
          },
          "pets_allowed": {
            "type": "boolean",
            "title": "Pets Allowed",
            "default": false
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoInput"
            },
            "type": "array",
            "title": "Photos"
          },
          "amenity_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Amenity Ids"
          },
          "price_per_night": {
            "type": "integer",
            "title": "Price Per Night"
          },
          "cleaning_fee": {
            "type": "integer",
            "title": "Cleaning Fee",
            "default": 0
          },
          "min_nights": {
            "type": "integer",
            "title": "Min Nights",
            "default": 1
          },
          "max_nights": {
            "type": "integer",
            "title": "Max Nights",
            "default": 30
          },
          "check_in_time": {
            "type": "string",
            "title": "Check In Time",
            "default": "15:00"
          },
          "check_out_time": {
            "type": "string",
            "title": "Check Out Time",
            "default": "11:00"
          },
          "house_rules": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "House Rules"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "category",
          "property_type",
          "room_type",
          "city",
          "country",
          "latitude",
          "longitude",
          "max_guests",
          "bedrooms",
          "beds",
          "baths",
          "photos",
          "amenity_ids",
          "price_per_night"
        ],
        "title": "CreateListingRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HostListingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "price_per_night": {
            "type": "integer",
            "title": "Price Per Night"
          },
          "rating_avg": {
            "type": "number",
            "title": "Rating Avg"
          },
          "rating_count": {
            "type": "integer",
            "title": "Rating Count"
          },
          "photos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photos"
          },
          "upcoming_bookings": {
            "type": "integer",
            "title": "Upcoming Bookings"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "city",
          "country",
          "price_per_night",
          "rating_avg",
          "rating_count",
          "photos",
          "upcoming_bookings",
          "is_active"
        ],
        "title": "HostListingResponse"
      },
...
      "PhotoInput": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "alt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt"
          }
        },
        "type": "object",
        "required": [
          "url",
          "position"
        ],
        "title": "PhotoInput"
      },
...
      "UpdateListingRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "property_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Type"
          },
          "room_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Type"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "max_guests": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Guests"
          },
          "bedrooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms"
          },
          "beds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beds"
          },
          "baths": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baths"
          },
          "pets_allowed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pets Allowed"
          },
          "photos": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PhotoInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photos"
          },
          "amenity_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenity Ids"
          },
          "price_per_night": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Night"
          },
          "cleaning_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cleaning Fee"
          },
          "min_nights": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Nights"
          },
          "max_nights": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Nights"
          },
          "check_in_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check In Time"
          },
          "check_out_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Out Time"
          },
          "house_rules": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "House Rules"
          }
        },
        "type": "object",
        "title": "UpdateListingRequest",
        "description": "All fields optional for PATCH/PUT."
      },
...
      "UploadResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "url",
          "filename"
        ],
        "title": "UploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}